feat(home): show last 3 articles
